Split demo project into multiple files
