+++ include/defines.h
@@ -1,0 +1,9 @@
+#ifndef __DEFINES_H__
+#define __DEFINES_H__
+
+#define mainGENERIC_PRIORITY (tskIDLE_PRIORITY)
+#define mainGENERIC_STACK_SIZE ((unsigned short)2560)
+
+#define FPS_FONT "IBMPlexSans-Bold.ttf"
+
+#endif // __DEFINES_H__
