+++ src/main.c
@@ -2,11 +2,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include <SDL2/SDL_scancode.h>
-
-#include "FreeRTOS.h"
+#include "main.h"
 #include "queue.h"
-#include "semphr.h"
 #include "task.h"
 
 #include "TUM_Ball.h"
@@ -20,170 +17,26 @@
 
 #include "AsyncIO.h"
 
-#define mainGENERIC_PRIORITY (tskIDLE_PRIORITY)
-#define mainGENERIC_STACK_SIZE ((unsigned short)2560)
+#include "defines.h"
+#include "demo_tasks.h"
+#include "async_sockets.h"
+#include "async_message_queues.h"
+#include "draw.h"
+#include "buttons.h"
 
-#define STATE_QUEUE_LENGTH 1
-
-#define STATE_COUNT 2
-
-#define STATE_ONE 0
-#define STATE_TWO 1
-
-#define NEXT_TASK 0
-#define PREV_TASK 1
-
-#define STARTING_STATE STATE_ONE
-
-#define STATE_DEBOUNCE_DELAY 300
-
-#define KEYCODE(CHAR) SDL_SCANCODE_##CHAR
-#define CAVE_SIZE_X SCREEN_WIDTH / 2
-#define CAVE_SIZE_Y SCREEN_HEIGHT / 2
-#define CAVE_X CAVE_SIZE_X / 2
-#define CAVE_Y CAVE_SIZE_Y / 2
-#define CAVE_THICKNESS 25
-#define LOGO_FILENAME "freertos.jpg"
-#define UDP_BUFFER_SIZE 2000
-#define UDP_TEST_PORT_1 1234
-#define UDP_TEST_PORT_2 4321
-#define MSG_QUEUE_BUFFER_SIZE 1000
-#define MSG_QUEUE_MAX_MSG_COUNT 10
-#define TCP_BUFFER_SIZE 2000
-#define TCP_TEST_PORT 2222
 
 #ifdef TRACE_FUNCTIONS
 #include "tracer.h"
 #endif
-
-static char *mq_one_name = "FreeRTOS_MQ_one_1";
-static char *mq_two_name = "FreeRTOS_MQ_two_1";
-aIO_handle_t mq_one = NULL;
-aIO_handle_t mq_two = NULL;
-aIO_handle_t udp_soc_one = NULL;
-aIO_handle_t udp_soc_two = NULL;
-aIO_handle_t tcp_soc = NULL;
 
 const unsigned char next_state_signal = NEXT_TASK;
 const unsigned char prev_state_signal = PREV_TASK;
 
 static TaskHandle_t StateMachine = NULL;
 static TaskHandle_t BufferSwap = NULL;
-static TaskHandle_t DemoTask1 = NULL;
-static TaskHandle_t DemoTask2 = NULL;
-static TaskHandle_t UDPDemoTask = NULL;
-static TaskHandle_t TCPDemoTask = NULL;
-static TaskHandle_t MQDemoTask = NULL;
-static TaskHandle_t DemoSendTask = NULL;
 
-static QueueHandle_t StateQueue = NULL;
-static SemaphoreHandle_t DrawSignal = NULL;
 
-static image_handle_t logo_image = NULL;
-
-typedef struct buttons_buffer {
-    unsigned char buttons[SDL_NUM_SCANCODES];
-    SemaphoreHandle_t lock;
-} buttons_buffer_t;
-
-static buttons_buffer_t buttons = { 0 };
-
-void checkDraw(unsigned char status, const char *msg)
-{
-    if (status) {
-        if (msg)
-            fprints(stderr, "[ERROR] %s, %s\n", msg,
-                    tumGetErrorMessage());
-        else {
-            fprints(stderr, "[ERROR] %s\n", tumGetErrorMessage());
-        }
-    }
-}
-
-/*
- * Changes the state, either forwards of backwards
- */
-void changeState(volatile unsigned char *state, unsigned char forwards)
-{
-    switch (forwards) {
-        case NEXT_TASK:
-            if (*state == STATE_COUNT - 1) {
-                *state = 0;
-            }
-            else {
-                (*state)++;
-            }
-            break;
-        case PREV_TASK:
-            if (*state == 0) {
-                *state = STATE_COUNT - 1;
-            }
-            else {
-                (*state)--;
-            }
-            break;
-        default:
-            break;
-    }
-}
-
-/*
- * Example basic state machine with sequential states
- */
-void basicSequentialStateMachine(void *pvParameters)
-{
-    unsigned char current_state = STARTING_STATE; // Default state
-    unsigned char state_changed =
-        1; // Only re-evaluate state if it has changed
-    unsigned char input = 0;
-
-    const int state_change_period = STATE_DEBOUNCE_DELAY;
-
-    TickType_t last_change = xTaskGetTickCount();
-
-    while (1) {
-        if (state_changed) {
-            goto initial_state;
-        }
-
-        // Handle state machine input
-        if (StateQueue)
-            if (xQueueReceive(StateQueue, &input, portMAX_DELAY) ==
-                pdTRUE)
-                if (xTaskGetTickCount() - last_change >
-                    state_change_period) {
-                    changeState(&current_state, input);
-                    state_changed = 1;
-                    last_change = xTaskGetTickCount();
-                }
-
-initial_state:
-        // Handle current state
-        if (state_changed) {
-            switch (current_state) {
-                case STATE_ONE:
-                    if (DemoTask2) {
-                        vTaskSuspend(DemoTask2);
-                    }
-                    if (DemoTask1) {
-                        vTaskResume(DemoTask1);
-                    }
-                    break;
-                case STATE_TWO:
-                    if (DemoTask1) {
-                        vTaskSuspend(DemoTask1);
-                    }
-                    if (DemoTask2) {
-                        vTaskResume(DemoTask2);
-                    }
-                    break;
-                default:
-                    break;
-            }
-            state_changed = 0;
-        }
-    }
-}
+SemaphoreHandle_t DrawSignal = NULL;
 
 void vSwapBuffers(void *pvParameters)
 {
@@ -199,471 +52,6 @@ void vSwapBuffers(void *pvParameters)
                         pdMS_TO_TICKS(frameratePeriod));
     }
 }
-
-void xGetButtonInput(void)
-{
-    if (xSemaphoreTake(buttons.lock, 0) == pdTRUE) {
-        xQueueReceive(buttonInputQueue, &buttons.buttons, 0);
-        xSemaphoreGive(buttons.lock);
-    }
-}
-
-void vDrawCaveBoundingBox(void)
-{
-    checkDraw(tumDrawFilledBox(CAVE_X - CAVE_THICKNESS,
-                               CAVE_Y - CAVE_THICKNESS,
-                               CAVE_SIZE_X + CAVE_THICKNESS * 2,
-                               CAVE_SIZE_Y + CAVE_THICKNESS * 2, TUMBlue),
-              __FUNCTION__);
-
-    checkDraw(tumDrawFilledBox(CAVE_X, CAVE_Y, CAVE_SIZE_X, CAVE_SIZE_Y,
-                               Aqua),
-              __FUNCTION__);
-}
-
-void vDrawCave(unsigned char ball_color_inverted)
-{
-    static unsigned short circlePositionX, circlePositionY;
-
-    vDrawCaveBoundingBox();
-
-    circlePositionX = CAVE_X + tumEventGetMouseX() / 2;
-    circlePositionY = CAVE_Y + tumEventGetMouseY() / 2;
-
-    if (ball_color_inverted)
-        checkDraw(tumDrawCircle(circlePositionX, circlePositionY, 20,
-                                Black),
-                  __FUNCTION__);
-    else
-        checkDraw(tumDrawCircle(circlePositionX, circlePositionY, 20,
-                                Silver),
-                  __FUNCTION__);
-}
-
-void vDrawHelpText(void)
-{
-    static char str[100] = { 0 };
-    static int text_width;
-    ssize_t prev_font_size = tumFontGetCurFontSize();
-
-    tumFontSetSize((ssize_t)30);
-
-    sprintf(str, "[Q]uit, [C]hange State");
-
-    if (!tumGetTextSize((char *)str, &text_width, NULL))
-        checkDraw(tumDrawText(str, SCREEN_WIDTH - text_width - 10,
-                              DEFAULT_FONT_SIZE * 0.5, Black),
-                  __FUNCTION__);
-
-    tumFontSetSize(prev_font_size);
-}
-
-#define FPS_AVERAGE_COUNT 50
-#define FPS_FONT "IBMPlexSans-Bold.ttf"
-
-void vDrawFPS(void)
-{
-    static unsigned int periods[FPS_AVERAGE_COUNT] = { 0 };
-    static unsigned int periods_total = 0;
-    static unsigned int index = 0;
-    static unsigned int average_count = 0;
-    static TickType_t xLastWakeTime = 0, prevWakeTime = 0;
-    static char str[10] = { 0 };
-    static int text_width;
-    int fps = 0;
-    font_handle_t cur_font = tumFontGetCurFontHandle();
-
-    if (average_count < FPS_AVERAGE_COUNT) {
-        average_count++;
-    }
-    else {
-        periods_total -= periods[index];
-    }
-
-    xLastWakeTime = xTaskGetTickCount();
-
-    if (prevWakeTime != xLastWakeTime) {
-        periods[index] =
-            configTICK_RATE_HZ / (xLastWakeTime - prevWakeTime);
-        prevWakeTime = xLastWakeTime;
-    }
-    else {
-        periods[index] = 0;
-    }
-
-    periods_total += periods[index];
-
-    if (index == (FPS_AVERAGE_COUNT - 1)) {
-        index = 0;
-    }
-    else {
-        index++;
-    }
-
-    fps = periods_total / average_count;
-
-    tumFontSelectFontFromName(FPS_FONT);
-
-    sprintf(str, "FPS: %2d", fps);
-
-    if (!tumGetTextSize((char *)str, &text_width, NULL))
-        checkDraw(tumDrawText(str, SCREEN_WIDTH - text_width - 10,
-                              SCREEN_HEIGHT - DEFAULT_FONT_SIZE * 1.5,
-                              Skyblue),
-                  __FUNCTION__);
-
-    tumFontSelectFontFromHandle(cur_font);
-    tumFontPutFontHandle(cur_font);
-}
-
-void vDrawLogo(void)
-{
-    static int image_height;
-
-    if ((image_height = tumDrawGetLoadedImageHeight(logo_image)) != -1)
-        checkDraw(tumDrawLoadedImage(logo_image, 10,
-                                     SCREEN_HEIGHT - 10 - image_height),
-                  __FUNCTION__);
-    else {
-        fprints(stderr,
-                "Failed to get size of image '%s', does it exist?\n",
-                LOGO_FILENAME);
-    }
-}
-
-void vDrawStaticItems(void)
-{
-    vDrawHelpText();
-    vDrawLogo();
-}
-
-void vDrawButtonText(void)
-{
-    static char str[100] = { 0 };
-
-    sprintf(str, "Axis 1: %5d | Axis 2: %5d", tumEventGetMouseX(),
-            tumEventGetMouseY());
-
-    checkDraw(tumDrawText(str, 10, DEFAULT_FONT_SIZE * 0.5, Black),
-              __FUNCTION__);
-
-    if (xSemaphoreTake(buttons.lock, 0) == pdTRUE) {
-        sprintf(str, "W: %d | S: %d | A: %d | D: %d",
-                buttons.buttons[KEYCODE(W)],
-                buttons.buttons[KEYCODE(S)],
-                buttons.buttons[KEYCODE(A)],
-                buttons.buttons[KEYCODE(D)]);
-        xSemaphoreGive(buttons.lock);
-        checkDraw(tumDrawText(str, 10, DEFAULT_FONT_SIZE * 2, Black),
-                  __FUNCTION__);
-    }
-
-    if (xSemaphoreTake(buttons.lock, 0) == pdTRUE) {
-        sprintf(str, "UP: %d | DOWN: %d | LEFT: %d | RIGHT: %d",
-                buttons.buttons[KEYCODE(UP)],
-                buttons.buttons[KEYCODE(DOWN)],
-                buttons.buttons[KEYCODE(LEFT)],
-                buttons.buttons[KEYCODE(RIGHT)]);
-        xSemaphoreGive(buttons.lock);
-        checkDraw(tumDrawText(str, 10, DEFAULT_FONT_SIZE * 3.5, Black),
-                  __FUNCTION__);
-    }
-}
-
-static int vCheckStateInput(void)
-{
-    if (xSemaphoreTake(buttons.lock, 0) == pdTRUE) {
-        if (buttons.buttons[KEYCODE(C)]) {
-            buttons.buttons[KEYCODE(C)] = 0;
-            if (StateQueue) {
-                xSemaphoreGive(buttons.lock);
-                xQueueSend(StateQueue, &next_state_signal, 0);
-                return 0;
-            }
-            return -1;
-        }
-        xSemaphoreGive(buttons.lock);
-    }
-
-    return 0;
-}
-
-void UDPHandlerOne(size_t read_size, char *buffer, void *args)
-{
-    prints("UDP Recv in first handler: %s\n", buffer);
-}
-
-void UDPHandlerTwo(size_t read_size, char *buffer, void *args)
-{
-    prints("UDP Recv in second handler: %s\n", buffer);
-}
-
-void vUDPDemoTask(void *pvParameters)
-{
-    char *addr = NULL; // Loopback
-    in_port_t port = UDP_TEST_PORT_1;
-
-    udp_soc_one = aIOOpenUDPSocket(addr, port, UDP_BUFFER_SIZE,
-                                   UDPHandlerOne, NULL);
-
-    prints("UDP socket opened on port %d\n", port);
-    prints("Demo UDP Socket can be tested using\n");
-    prints("*** netcat -vv localhost %d -u ***\n", port);
-
-    port = UDP_TEST_PORT_2;
-
-    udp_soc_two = aIOOpenUDPSocket(addr, port, UDP_BUFFER_SIZE,
-                                   UDPHandlerTwo, NULL);
-
-    prints("UDP socket opened on port %d\n", port);
-    prints("Demo UDP Socket can be tested using\n");
-    prints("*** netcat -vv localhost %d -u ***\n", port);
-
-    while (1) {
-        vTaskDelay(pdMS_TO_TICKS(1000));
-    }
-}
-
-void MQHandlerOne(size_t read_size, char *buffer, void *args)
-{
-    prints("MQ Recv in first handler: %s\n", buffer);
-}
-
-void MQHanderTwo(size_t read_size, char *buffer, void *args)
-{
-    prints("MQ Recv in second handler: %s\n", buffer);
-}
-
-void vDemoSendTask(void *pvParameters)
-{
-    static char *test_str_1 = "UDP test 1";
-    static char *test_str_2 = "UDP test 2";
-    static char *test_str_3 = "TCP test";
-
-    while (1) {
-        prints("*****TICK******\n");
-        if (mq_one) {
-            aIOMessageQueuePut(mq_one_name, "Hello MQ one");
-        }
-        if (mq_two) {
-            aIOMessageQueuePut(mq_two_name, "Hello MQ two");
-        }
-
-        if (udp_soc_one)
-            aIOSocketPut(UDP, NULL, UDP_TEST_PORT_1, test_str_1,
-                         strlen(test_str_1));
-        if (udp_soc_two)
-            aIOSocketPut(UDP, NULL, UDP_TEST_PORT_2, test_str_2,
-                         strlen(test_str_2));
-        if (tcp_soc)
-            aIOSocketPut(TCP, NULL, TCP_TEST_PORT, test_str_3,
-                         strlen(test_str_3));
-
-        vTaskDelay(pdMS_TO_TICKS(1000));
-    }
-}
-
-void vMQDemoTask(void *pvParameters)
-{
-    mq_one = aIOOpenMessageQueue(mq_one_name, MSG_QUEUE_MAX_MSG_COUNT,
-                                 MSG_QUEUE_BUFFER_SIZE, MQHandlerOne, NULL);
-    mq_two = aIOOpenMessageQueue(mq_two_name, MSG_QUEUE_MAX_MSG_COUNT,
-                                 MSG_QUEUE_BUFFER_SIZE, MQHanderTwo, NULL);
-
-    while (1)
-
-    {
-        vTaskDelay(pdMS_TO_TICKS(1000));
-    }
-}
-
-void TCPHandler(size_t read_size, char *buffer, void *args)
-{
-    prints("TCP Recv: %s\n", buffer);
-}
-
-void vTCPDemoTask(void *pvParameters)
-{
-    char *addr = NULL; // Loopback
-    in_port_t port = TCP_TEST_PORT;
-
-    tcp_soc =
-        aIOOpenTCPSocket(addr, port, TCP_BUFFER_SIZE, TCPHandler, NULL);
-
-    prints("TCP socket opened on port %d\n", port);
-    prints("Demo TCP socket can be tested using\n");
-    prints("*** netcat -vv localhost %d ***\n", port);
-
-    while (1) {
-        vTaskDelay(10);
-    }
-}
-
-void vDemoTask1(void *pvParameters)
-{
-    char *ball_spritesheet_path = tumUtilFindResourcePath("ball_spritesheet.png");
-    image_handle_t ball_spritesheet_image =
-        tumDrawLoadImage(ball_spritesheet_path);
-    spritesheet_handle_t ball_spritesheet =
-        tumDrawLoadSpritesheet(ball_spritesheet_image, 25, 1);
-    animation_handle_t ball_animation =
-        tumDrawAnimationCreate(ball_spritesheet);
-    tumDrawAnimationAddSequence(ball_animation, "FORWARDS", 0, 0,
-                                SPRITE_SEQUENCE_HORIZONTAL_POS, 24);
-    tumDrawAnimationAddSequence(ball_animation, "REVERSE", 0, 23,
-                                SPRITE_SEQUENCE_HORIZONTAL_NEG, 24);
-    sequence_handle_t forward_sequence =
-        tumDrawAnimationSequenceInstantiate(ball_animation, "FORWARDS",
-                                            40);
-    sequence_handle_t reverse_sequence =
-        tumDrawAnimationSequenceInstantiate(ball_animation, "REVERSE",
-                                            40);
-    TickType_t xLastFrameTime = xTaskGetTickCount();
-
-    while (1) {
-        if (DrawSignal)
-            if (xSemaphoreTake(DrawSignal, portMAX_DELAY) ==
-                pdTRUE) {
-                tumEventFetchEvents(FETCH_EVENT_BLOCK |
-                                    FETCH_EVENT_NO_GL_CHECK);
-                xGetButtonInput(); // Update global input
-
-                // Clear screen
-                checkDraw(tumDrawClear(White), __FUNCTION__);
-                vDrawStaticItems();
-                vDrawCave(tumEventGetMouseLeft());
-                vDrawButtonText();
-                tumDrawAnimationDrawFrame(
-                    forward_sequence,
-                    xTaskGetTickCount() - xLastFrameTime,
-                    SCREEN_WIDTH - 50, SCREEN_HEIGHT - 60);
-                tumDrawAnimationDrawFrame(
-                    reverse_sequence,
-                    xTaskGetTickCount() - xLastFrameTime,
-                    SCREEN_WIDTH - 90,
-                    SCREEN_HEIGHT - 60);
-                checkDraw(tumDrawSprite(ball_spritesheet, 5, 0,
-                                        SCREEN_WIDTH - 130, SCREEN_HEIGHT - 60),
-                          __FUNCTION__);
-
-                xLastFrameTime = xTaskGetTickCount();
-
-                // Draw FPS in lower right corner
-                vDrawFPS();
-
-                // Get input and check for state change
-                vCheckStateInput();
-            }
-    }
-}
-
-void playBallSound(void *args)
-{
-    tumSoundPlaySample(a3);
-}
-
-void vDemoTask2(void *pvParameters)
-{
-    TickType_t xLastWakeTime, prevWakeTime;
-    xLastWakeTime = xTaskGetTickCount();
-    prevWakeTime = xLastWakeTime;
-
-    ball_t *my_ball = createBall(SCREEN_WIDTH / 2, SCREEN_HEIGHT / 2, Black,
-                                 20, 1000, &playBallSound, NULL);
-    setBallSpeed(my_ball, 250, 250, 0, SET_BALL_SPEED_AXES);
-
-    // Left wall
-    wall_t *left_wall =
-        createWall(CAVE_X - CAVE_THICKNESS, CAVE_Y, CAVE_THICKNESS,
-                   CAVE_SIZE_Y, 0.2, Red, NULL, NULL);
-    // Right wall
-    wall_t *right_wall =
-        createWall(CAVE_X + CAVE_SIZE_X, CAVE_Y, CAVE_THICKNESS,
-                   CAVE_SIZE_Y, 0.2, Red, NULL, NULL);
-    // Top wall
-    wall_t *top_wall =
-        createWall(CAVE_X - CAVE_THICKNESS, CAVE_Y - CAVE_THICKNESS,
-                   CAVE_SIZE_X + CAVE_THICKNESS * 2, CAVE_THICKNESS,
-                   0.2, Blue, NULL, NULL);
-    // Bottom wall
-    wall_t *bottom_wall =
-        createWall(CAVE_X - CAVE_THICKNESS, CAVE_Y + CAVE_SIZE_Y,
-                   CAVE_SIZE_X + CAVE_THICKNESS * 2, CAVE_THICKNESS,
-                   0.2, Blue, NULL, NULL);
-    unsigned char collisions = 0;
-
-    prints("Task 1 init'd\n");
-
-    while (1) {
-        if (DrawSignal)
-            if (xSemaphoreTake(DrawSignal, portMAX_DELAY) ==
-                pdTRUE) {
-                xLastWakeTime = xTaskGetTickCount();
-
-                xGetButtonInput(); // Update global button data
-
-                // Clear screen
-                checkDraw(tumDrawClear(White), __FUNCTION__);
-
-                vDrawStaticItems();
-
-                // Draw the walls
-                checkDraw(tumDrawFilledBox(
-                              left_wall->x1, left_wall->y1,
-                              left_wall->w, left_wall->h,
-                              left_wall->colour),
-                          __FUNCTION__);
-                checkDraw(tumDrawFilledBox(right_wall->x1,
-                                           right_wall->y1,
-                                           right_wall->w,
-                                           right_wall->h,
-                                           right_wall->colour),
-                          __FUNCTION__);
-                checkDraw(tumDrawFilledBox(
-                              top_wall->x1, top_wall->y1,
-                              top_wall->w, top_wall->h,
-                              top_wall->colour),
-                          __FUNCTION__);
-                checkDraw(tumDrawFilledBox(bottom_wall->x1,
-                                           bottom_wall->y1,
-                                           bottom_wall->w,
-                                           bottom_wall->h,
-                                           bottom_wall->colour),
-                          __FUNCTION__);
-
-                // Check if ball has made a collision
-                collisions = checkBallCollisions(my_ball, NULL,
-                                                 NULL);
-                if (collisions) {
-                    prints("Collision\n");
-                }
-
-                // Update the balls position now that possible collisions have
-                // updated its speeds
-                updateBallPosition(
-                    my_ball, xLastWakeTime - prevWakeTime);
-
-                // Draw the ball
-                checkDraw(tumDrawCircle(my_ball->x, my_ball->y,
-                                        my_ball->radius,
-                                        my_ball->colour),
-                          __FUNCTION__);
-
-                // Draw FPS in lower right corner
-                vDrawFPS();
-
-                // Check for state change
-                vCheckStateInput();
-
-                // Keep track of when task last ran so that you know how many ticks
-                //(in our case miliseconds) have passed so that the balls position
-                // can be updated appropriatley
-                prevWakeTime = xLastWakeTime;
-            }
-    }
-}
-
-#define PRINT_TASK_ERROR(task) PRINT_ERROR("Failed to print task ##task");
 
 int main(int argc, char *argv[])
 {
@@ -706,16 +94,14 @@ int main(int argc, char *argv[])
         goto err_init_safe_print;
     }
 
-    logo_image = tumDrawLoadImage(LOGO_FILENAME);
 
     atexit(aIODeinit);
 
     //Load a second font for fun
     tumFontLoadFont(FPS_FONT, DEFAULT_FONT_SIZE);
 
-    buttons.lock = xSemaphoreCreateMutex(); // Locking mechanism
-    if (!buttons.lock) {
-        PRINT_ERROR("Failed to create buttons lock");
+    if (buttonsInit()) {
+        PRINT_ERROR("Failed to init buttons");
         goto err_buttons_lock;
     }
 
@@ -746,31 +132,19 @@ int main(int argc, char *argv[])
     }
 
     /** Demo Tasks */
-    if (xTaskCreate(vDemoTask1, "DemoTask1", mainGENERIC_STACK_SIZE * 2,
-                    NULL, mainGENERIC_PRIORITY + 1, &DemoTask1) != pdPASS) {
-        PRINT_TASK_ERROR("DemoTask1");
-        goto err_demotask1;
-    }
-    if (xTaskCreate(vDemoTask2, "DemoTask2", mainGENERIC_STACK_SIZE * 2,
-                    NULL, mainGENERIC_PRIORITY + 1, &DemoTask2) != pdPASS) {
-        PRINT_TASK_ERROR("DemoTask2");
-        goto err_demotask2;
+    if (createDemoTasks()) {
+        goto err_demotasks;
     }
 
     /** SOCKETS */
-    xTaskCreate(vUDPDemoTask, "UDPTask", mainGENERIC_STACK_SIZE * 2, NULL,
-                configMAX_PRIORITIES - 1, &UDPDemoTask);
-    xTaskCreate(vTCPDemoTask, "TCPTask", mainGENERIC_STACK_SIZE, NULL,
-                configMAX_PRIORITIES - 1, &TCPDemoTask);
+    if (createSocketTasks()) {
+        goto err_sockettasks;
+    }
 
     /** POSIX MESSAGE QUEUES */
-    xTaskCreate(vMQDemoTask, "MQTask", mainGENERIC_STACK_SIZE * 2, NULL,
-                configMAX_PRIORITIES - 1, &MQDemoTask);
-    xTaskCreate(vDemoSendTask, "SendTask", mainGENERIC_STACK_SIZE * 2, NULL,
-                configMAX_PRIORITIES - 1, &DemoSendTask);
-
-    vTaskSuspend(DemoTask1);
-    vTaskSuspend(DemoTask2);
+    if (createMessageQueueTasks()) {
+        goto err_messagequeuetasks;
+    }
 
     tumFUtilPrintTaskStateList();
 
@@ -778,9 +152,11 @@ int main(int argc, char *argv[])
 
     return EXIT_SUCCESS;
 
-err_demotask2:
-    vTaskDelete(DemoTask1);
-err_demotask1:
+err_messagequeuetasks:
+    deleteSocketTasks();
+err_sockettasks:
+    deleteDemoTasks();
+err_demotasks:
     vTaskDelete(BufferSwap);
 err_bufferswap:
     vTaskDelete(StateMachine);
@@ -789,7 +165,7 @@ err_statemachine:
 err_state_queue:
     vSemaphoreDelete(DrawSignal);
 err_draw_signal:
-    vSemaphoreDelete(buttons.lock);
+    buttonsExit();
 err_buttons_lock:
     tumSoundExit();
 err_init_audio:
